Add concurrency and memoization
